+++ ts/components/Dialog/AlertDialog.tsx
@@ -1,8 +1,9 @@
-import React, { useEffect } from 'react';
+import React, { useRef, useState, useEffect } from 'react';
 import { Portal } from '../Portal';
 import { DialogBackdrop } from './DialogBackdrop';
 import { useDialog } from './useDialog';
 import { DialogContext } from './DialogContext';
+import { CSSTransition } from 'react-transition-group';
 
 interface AlertDialogProps {
   /** Content for the dialog. */
@@ -11,6 +12,10 @@ interface AlertDialogProps {
   closeDialog: () => void;
   /** Controls whether the dialog is open or not. */
   open: boolean;
+  /** Controls whether the dialog animates in. */
+  animateIn?: boolean;
+  /** Controls whether the dialog animates out. */
+  animateOut?: boolean;
   /** Adds class names to the backdrop element. */
   backdropClassName?: string;
   /** Adds class names to the dialog wrapper element. */
@@ -26,6 +31,8 @@ interface AlertDialogProps {
 }
 
 export const AlertDialog = ({
+  animateIn = true,
+  animateOut = true,
   backdropClassName,
   children,
   closeDialog,
@@ -49,6 +56,8 @@ export const AlertDialog = ({
     ariaLabelSelector,
     ariaDescriptionSelector,
   } = useDialog({
+    animationDirection: `up`,
+    animationDistance: `md`,
     backdrop: true,
     closeDialog,
     dialogClassName,
@@ -65,17 +74,38 @@ export const AlertDialog = ({
     leastDestructiveRef?.current?.focus();
   }, [leastDestructiveRef]);
 
+  const timeout = 250;
+  const nodeRef = useRef(null);
+  const [visible, setVisible] = useState(false);
+
+  useEffect(() => {
+    if (open) {
+      setVisible(true);
+    }
+  }, [open]);
+
   return (
     <>
       {open ? (
         <DialogContext.Provider value={{ ariaLabelSelector, ariaDescriptionSelector, closeDialog }}>
           <Portal className="h-dialog-portal">
-            <div {...getDialogRootProps()}>
-              <div {...getDialogContainerProps()}>
-                <DialogBackdrop className={backdropClassName} closeDialog={() => {}} />
-                <div {...getDialogProps()}>{children}</div>
+            <CSSTransition
+              nodeRef={nodeRef}
+              in={open && visible}
+              appear={animateIn || animateOut}
+              timeout={timeout}
+              enter={animateIn}
+              exit={animateOut}
+              classNames="h-transition-"
+              onExited={() => setVisible(false)}
+            >
+              <div {...getDialogRootProps()} ref={nodeRef}>
+                <div {...getDialogContainerProps()}>
+                  <DialogBackdrop className={backdropClassName} closeDialog={() => { }} />
+                  <div {...getDialogProps()}>{children}</div>
+                </div>
               </div>
-            </div>
+            </CSSTransition>
           </Portal>
         </DialogContext.Provider>
       ) : null}